+++ src/chat21-core/providers/mqtt/mqtt-conversation-handler.ts
@@ -99,7 +99,6 @@ export class MQTTConversationHandler extends ConversationHandlerService {
         }
         this.chat21Service.chatClient.lastMessages(this.conversationWith, (err, messages) => {
             if (!err) {
-                this.logger.log('[MQTTConversationHandlerSERVICE] message lastMessages:', messages);
                 messages.sort(compareValues('timestamp', 'asc'));
                 messages.forEach(message => {
                     const msg: MessageModel = message;