+++ src/chat21-core/utils/utils-message.ts
@@ -27,10 +27,10 @@ export function isFrame(message: any) {
 
 /** */
 export function isFile(message: any) {
-    if (message && message.type && message.type === 'file' && message.metadata && message.metadata.src && !message.metadata.type.includes('audio')) {
-      return true;
-    }
-    return false;
+  if (message && message.type && message.type === 'file' && message.metadata && message.metadata.src && !message.metadata.type.includes('audio')) {
+    return true;
+  }
+  return false;
 }
 
 export function isAudio(message: any) {