+++ src/chat21-core/utils/utils.ts
@@ -654,22 +654,6 @@ export function searchEmailOrUrlInString(item) {
   }
 }
 
-
-export function isURL(str: string) {
-  var pattern = new RegExp('^(https?:\/\/)?' + // protocol
-    '((([a-z\d]([a-z\d-]*[a-z\d])*)\.)+[a-z]{2,}|' + // domain name
-    '((\d{1,3}\.){3}\d{1,3}))' + // OR ip (v4) address
-    '(\:\d+)?(\/[-a-z\d%_.~+]*)*' + // port and path
-    '(\?[;&a-z\d%_.~+=-]*)?' + // query string
-    '(\#[-a-z\d_]*)?$', 'i'); // fragment locater
-  // console.log('pattern: ', pattern);
-  if (!pattern.test(str)) {
-    return false;
-  } else {
-    return true;
-  }
-}
-
 export function convertColorToRGBA(color, opacity) {
   let result = color;
   // console.log('convertColorToRGBA' + color, opacity);